Separate configuration from logic
