+++ docs/index.html
@@ -27,8 +27,9 @@
       <li><code>node legendaly.js</code> で起動します。</li>
     </ol>
     <p>
-      <code>.env</code> ファイルまたは環境変数で
-      <code>TONE</code>、<code>FETCH_INTERVAL</code>、<code>COLOR_TONE</code> などを設定できます。
+      設定は <code>config.js</code> で集中管理されています。
+      <code>.env</code> ファイルや環境変数で
+      <code>TONE</code>、<code>FETCH_INTERVAL</code> などを上書きできます。
     </p>
   </section>
   <footer>
